+++ pyincore_viz/__init__.py
@@ -1,3 +1,4 @@
 from pyincore_viz.plotutil import PlotUtil
 from pyincore_viz.geoutil import GeoUtil
+from pyincore_viz.csvmaputil import CsvMapUtil
 import pyincore_viz.globals