+++ docs/source/conf.py
@@ -31,9 +31,9 @@ project = 'pyIncore_viz'
 author = ''
 
 # The short X.Y version
-version = '0.2.2'
+version = '0.2.3'
 # The full version, including alpha/beta/rc tags
-release = '0.2.2'
+release = '0.2.3'
 
 
 # -- General configuration ---------------------------------------------------