+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- CSV file mapping tool using the csv files in given folder. [INCORE1-506](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-506)
+
 ## [0.2.2] - 2020-07-31
 ### Added
 - PEP-8 format test to the library.  [INCORE1-706](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-706)